\name{internals}
\alias{Moran}
\alias{VPcalc}
\alias{forcito}

\title{Las funciones Internas}
\description{Internas para spfod}
\usage{
Moran(modelo, mtz, p)
VPcalc(mtz, OD, Mat)
forcito(modelo, vp, datos, vpmod)
}
\arguments{
  \item{modelo}{En Moran y forcito es el mismo}
  \item{mtz}{En Moran y VPcalc es la misma ?verificar?}
  \item{p}{tres}

  \item{OD}{asaa}
  \item{Mat}{ssdfa}

  \item{vp}{fasdf}
  \item{datos}{asdf}
  \item{vpmod}{asf}
}

\details{Aca debemos explicar algo mas}
\value{salida
%%  \item{comp1 }{Description of 'comp1'}
}
\author{Eugenia Riano, Martin Hansz, Fernando Massa, Antonio Rey, Andres Castrillejo}

