\name{MoMoMo-package}
\alias{MoMoMo-package}
\alias{MoMoMo}
\docType{package}
\title{
\packageTitle{MoMoMo}
}
\description{
\packageDescription{MoMoMo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MoMoMo}
\packageIndices{MoMoMo}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MoMoMo}

Maintainer: \packageMaintainer{MoMoMo}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
