\name{tabla_empleo}
\alias{tabla_empleo}
\docType{data}
\title{Tabla de Empleos - Municipios}
\description{data.frame con el total de empleados por Municipo}
\usage{data("tabla_empleo")}
\format{
  A data frame with 8 observations on the following 3 variables.
  \describe{
    \item{\code{municipio}}{a factor with levels \code{A} \code{B} \code{C} \code{CH} \code{D} \code{E} \code{F} \code{G}}
    \item{\code{empleo}}{a numeric vector}
    \item{\code{prop}}{a numeric vector}
  }
}
\details{Algo del origen de los datos y eventualidades}
\examples{
data(tabla_empleo)
## maybe str(tabla_empleo) ; plot(tabla_empleo) ...
}
\keyword{datasets}
